Backend/llm/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

httpx==0.25.2

# LLM & Embeddings
openai==1.12.0

# Utilities
structlog==24.1.0
email-validator==2.1.0